requirements.txt
streamlit
google-genai
google-generativeai
python-dotenv
requests
pytest
markdown
